requirements.txt
numpy
pandas
openpyxl
pyarrow
flask
requests
influxdb-client
fastparquet
sqlalchemy
pyyaml

